+++ Source/Cancellable.swift
@@ -1,13 +1,13 @@
 /// Protocol to define the opaque type returned from a request
 public protocol Cancellable {
-    var cancelled: Bool { get }
+    var isCancelled: Bool { get }
     func cancel()
 }
 
 internal class CancellableWrapper: Cancellable {
     internal var innerCancellable: Cancellable = SimpleCancellable()
 
-    var cancelled: Bool { return innerCancellable.cancelled }
+    var isCancelled: Bool { return innerCancellable.isCancelled }
 
     internal func cancel() {
         innerCancellable.cancel()
@@ -15,8 +15,8 @@ internal class CancellableWrapper: Cancellable {
 }
 
 internal class SimpleCancellable: Cancellable {
-    var cancelled = false
+    var isCancelled = false
     func cancel() {
-        cancelled = true
+        isCancelled = true
     }
 }