+++ Changelog.md
@@ -1,4 +1,7 @@
 # Next
+
+- **Breaking Change** Renamed `cancelled` in the `Cancellable` protocol to `isCancelled`.
+- **Breaking Change** Renamed `URL` in `Endpoint` to `url`.
 - **Breaking Change** Renamed `StructTarget` to `MultiTarget`.
 - Demo project has been updated with new DemoMultiTarget target, new project
 structure and more.